Make EncoderFormatters fast instead of correct. :p
AbstractLoggingEventEncoderFormatter.isCompatible(object) is now simply checking if the given object contains a LoggingEvent.

It previously checked whether toString(object) returned a non-null value, which was quite insane…

This means that a “Copy event” action can now be active in the menu but won’t be able to actually format an event.

Given that this only happens in case of malformed events (like ones without a logger name) and that those checks are expensive and performed very often, I’d argue that this is a good move.

This also includes a fix regarding excludedIndices handling in AbstractBasicFormatterSpec.